requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# LangChain ecosystem
langchain==0.1.0
langgraph==0.0.62
langchain-community==0.0.10
langchain-core==0.1.0

# Database
pymilvus==2.3.4
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ML and LLM
torch>=2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
bitsandbytes==0.41.3
accelerate==0.25.0

# Web search and scraping
duckduckgo-search==3.9.6
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Reranking
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
huggingface-hub==0.19.4

# Production WSGI/ASGI server
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# API framework extensions
python-multipart==0.0.6